+++ packages/ui-core/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+- **Carousel**
+    - create new theme `lk`
+
 ## Release 1.0.16
 - **Select**
     - import fix